Move FF type to core utils file
