--- packages/core/src/featureFlags.ts
@@ -1,1 +1,0 @@
-export type FeatureFlag = { readonly flag: string; readonly result: boolean };
